test(rpc): consolidate ZoneRpcApi fixtures
